Moving tests into appropriate directories to match implementation
