+++ tests/geometry/chopper_test_helpers.py
