+++ tests/geometry/test_chopper_details.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pytest
 from nexus_constructor.geometry.disk_chopper.chopper_details import ChopperDetails
-from tests.chopper_test_helpers import (
+from tests.geometry.chopper_test_helpers import (
     N_SLITS,
     RADIUS_LENGTH,
     SLIT_HEIGHT_LENGTH,
