+++ tests/geometry/test_disk_chopper_checker.py
@@ -20,7 +20,7 @@ from nexus_constructor.geometry.disk_chopper.disk_chopper_checker import (
 )
 from nexus_constructor.model.dataset import Dataset
 from nexus_constructor.model.value_type import ValueTypes
-from tests.chopper_test_helpers import (  # noqa: F401
+from tests.geometry.chopper_test_helpers import (  # noqa: F401
     N_SLITS,
     DEGREES_EDGES_ARR,
     RADIUS_LENGTH,
