+++ tests/geometry/test_geometry_loader.py
