+++ tests/geometry/test_no_shape_geometry.py
