+++ tests/instrument_view/test_gnomon.py
