+++ tests/instrument_view/test_instrument_view.py
