+++ tests/instrument_view/test_instrument_view_axes.py
