+++ tests/instrument_view/test_neutron_animation_controller.py
