+++ tests/instrument_view/test_off_renderer.py
