+++ tests/instrument_view/test_qentity_utils.py
